refactor: Refactor org_custom_roles.go into multiple files (#3291)
